% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{bayes_factor.bmult}
\alias{bayes_factor.bmult}
\title{Extracts information about computed Bayes factors from object of class \code{bmult}}
\usage{
\method{bayes_factor}{bmult}(x, ...)
}
\arguments{
\item{x}{object of class \code{bmult_bridge} as returned from \code{multBayesInformed}}

\item{...}{further arguments, currently ignored}
}
\value{
Returns \code{list} with two dataframes from an object of class \code{bmult}. The first dataframe bf_all summarizes information
the Bayes factor for equality and inequality constraints. The second dataframe bf_ineq summarized information about the Bayes factor for inequality
constraints.
}
\description{
Extracts information about computed Bayes factors from object of class \code{bmult}
}
