% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{bayes_factor}
\alias{bayes_factor}
\title{S3 method for class 'bayes_factor.bmult'}
\usage{
bayes_factor(x)
}
\arguments{
\item{x}{object of class \code{bmult} as returned from \code{\link{mult_bf_informed}} or \code{\link{binom_bf_informed}}
or an object of class \code{bmult_bridge} as returned from \code{\link{mult_bf_inequality}} or \code{\link{binom_bf_inequality}}}
}
\value{
Returns \code{list} with two \code{data.frames}. The first dataframe \code{bf_table} summarizes information
the Bayes factor for equality and inequality constraints. The second dataframe \code{$bf_ineq_table} summarized
information about the Bayes factor for inequality constraints, that is, the log marginal likelihood estimates
for the constrained prior and posterior distribution.
}
\description{
Extracts information about computed Bayes factors from object of class \code{bmult}
}
\examples{
# data
x <- c(3, 4, 10, 11)
n <- c(15, 12, 12, 12)
# priors
a <- c(1, 1, 1, 1)
b <- c(1, 1, 1, 1)
# informed hypothesis
factor_levels <- c('theta1', 'theta2', 'theta3', 'theta4')
Hr            <- c('theta1', '<',  'theta2', '<', 'theta3', '<', 'theta4')

## Multinomial Case
out_mult  <- mult_bf_informed(x=x, Hr=Hr, a=a, factor_levels=factor_levels,
niter=1e3, seed=2020)
bayes_factor(out_mult)
}
