% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_binomBayesInformed.R
\name{binomBayesInformed}
\alias{binomBayesInformed}
\title{Evaluates Informed Hypotheses on Multiple Binomial Parameters}
\usage{
binomBayesInformed(
  factor_levels,
  Hr,
  a,
  b,
  counts,
  total,
  niter = 50000,
  bf_type = "LogBFer",
  seed = NULL
)
}
\arguments{
\item{factor_levels}{character. vector with categories}

\item{Hr}{character. vector encoding the user specified order restriction}

\item{a}{numeric. vector with concentration parameters}

\item{b}{numeric vector with beta parameters}

\item{counts}{numeric. vector with data}

\item{total}{numeric vector with total number of observations}

\item{niter}{numeric. vector with number of samples to be drawn}

\item{bf_type}{'bf_type'. The Bayes factor type can be either 'LogBFer', 'BFer', or 'BFre'. Default is 'LogBFer'}

\item{seed}{set the seed for version control.}
}
\value{
list consisting of the following elements:
(1) BF: Bayes factor for restricted hypothesis compared to the encompassing hypothesis
(2) restrictions: full restriction list
(4) logBFe_inequalities: log Bayes factor for inequality constrained parameters
}
\description{
Computes Bayes factor for inequality constrained binomial parameters using the encompassing prior approach.
Restricted hypothesis Hr states that binomial proportions follow a particular trend.
Alternative hypothesis He states that binomial proportions are free to vary.
}
\examples{
# data
counts <- c(3, 4, 10, 11)
total  <- c(15, 12, 12, 12)
# priors
a <- c(1, 1, 1, 1)
b <- c(1, 1, 1, 1)
# restricted hypothesis
factor_levels <- c('binom1', 'binom2', 'binom3', 'binom4')
Hr            <- c('binom1', '<',  'binom2', '<', 'binom3', '<', 'binom4')
output_total  <- binomBayesInformed(factor_levels, Hr, a, b, counts, total, niter = 5e3, bf_type = 'LogBFer', seed=2020)
}
\seealso{
Other evalInformed: 
\code{\link{multBayesInformed}()}
}
\concept{evalInformed}
