% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_inequalities_binom.R
\name{binomBfInequality}
\alias{binomBfInequality}
\title{Computes Bayes Factors For Inequality Constrained Inedepent Binomial Parameters}
\usage{
binomBfInequality(
  samples,
  restrictions,
  a = rep(1, ncol(samples)),
  b = rep(1, ncol(samples)),
  counts = NULL,
  total = NULL,
  prior = FALSE,
  index = 1,
  maxiter = 1000,
  seed = NULL,
  ...
)
}
\arguments{
\item{samples}{matrix of dimension (nsamples x nparams) with samples from independent truncated Binomial densities}

\item{restrictions}{either \code{character vector} containing the user specified order restriction or \code{list} of class \code{bmult_rl} as returned from \code{generateRestrictionList} that encodes
inequality constraints for each independent restriction.}

\item{a}{numeric vector with values of alpha parameters of the beta distribution}

\item{b}{numeric vector with values of beta parameters of the beta distribution}

\item{counts}{numeric vector with number of successes}

\item{total}{numeric vector with total number of observations}

\item{prior}{logical. If TRUE the function will ignore the data and sample from the prior distribution}

\item{index}{index of current restriction. Default is 1.}

\item{maxiter}{maximum number of iterations for the iterative updating scheme. Default is 1,000 to avoid infinite loops.}

\item{seed}{set the seed for version control.}

\item{...}{additional arguments (currently ignored).}
}
\value{
list consisting of the following elements:
(1) eval: list consisting of the following elements
q11: log posterior evaluations for posterior samples.
q12: log proposal evaluations for posterior samples.
q21: log posterior evaluations for samples from proposal.
q22: log proposal evaluations for samples from proposal.
(2) niter: number of iterations of the iterative updating scheme.
(3) logml: estimate of log marginal likelihood.
(4) hyp: character vector that contains the inequality constrained hypothesis
}
\description{
Computes Bayes factor for inequality constrained binomial parameters using a bridge sampling routine.
Restricted hypothesis Hr states that binomial proportions follow a particular trend.
Alternative hypothesis He states that binomial proportions are free to vary.
}
