% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_helper_functions.R
\name{.collapseCategories}
\alias{.collapseCategories}
\title{Collapse Categories That Are Constrained To Be Equal}
\usage{
.collapseCategories(
  counts,
  equalities,
  is_numeric_value = TRUE,
  correct = FALSE,
  adjusted_priors_for_equalities = NULL
)
}
\arguments{
\item{counts}{either a character string or a numeric vector to be collapsed.}

\item{equalities}{index of elements in counts that are equality constrained.}

\item{is_numeric_value}{logical. If TRUE, the input is a numeric vector and if FALSE the input is a character vector.}

\item{correct}{logical. If TRUE a correction for marginalization is applied to numeric vectors. A correction is needed,
only if the function collapses concentration parameters.}

\item{adjusted_priors_for_equalities}{is a list with containing the adjusted priors for equality constraints. The list should have}
}
\value{
collapsed character or numeric vector.
}
\description{
This function collapses categories that are constrained to be equal
It can collapse character vector (i.e., category names), as well as numeric vectors
(i.e., data and concentration parameters). If the input is a character vector, the first element within the equalities
is retained.
}
