% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{.computeRMSE}
\alias{.computeRMSE}
\title{Computes The Relative Mean-Squared Error For Marginal Likelihood Estimate}
\usage{
\code{
computeRMSE(bridge_object, ...)
## S3 method for class 'bridge'
computeRMSE(bridge_object, ...)
}
}
\arguments{
\item{bridge_output}{an object of class \code{"bridge"} as returned from multBayesInformed}

\item{...}{additional arguments (currently ignored).}
}
\value{
Returns a list with the following components:
\describe{
  \item{re2}{approximate relative mean-squared error for marginal likelihood estimate.}
  \item{cv}{approximate coefficient of variation for marginal likelihood estimate (assumes that bridge estimate is unbiased).}
  \item{percentage}{approximate percentage error of marginal likelihood estimate.}
}
}
\description{
Computes error measures for estimated marginal likelihood.
}
\details{
Computes error measures for marginal likelihood bridge sampling estimates. The approximate errors for bridge sampling estimates are based on Fruehwirth-Schnatter (2004). 
Code is based on \code{error_measures} function of the \code{R} package \code{bridgesampling}.
}
\references{
\insertRef{gronau2017bridgesampling}{multibridge} \insertRef{fruhwirth2004estimating}{multibridge}
}
\author{
Quentin F. Gronau
}
