% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_helper_functions.R
\name{.splitAt}
\alias{.splitAt}
\title{Split A Character Vector At A Given Symbol}
\usage{
.splitAt(vector, signs)
}
\arguments{
\item{vector}{character vector to be broken into separate components.}

\item{signs}{character string containing a regular expression that is an element in vector and serves as breaking point.}
}
\value{
a list with separate components
}
\description{
This function splits a character vector into separate components
using a specified breaking point.
}
