% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{generate_restriction_list.bmult}
\alias{generate_restriction_list.bmult}
\title{Extracts restriction list from an object of class \code{bmult}}
\usage{
generate_restriction_list.bmult(x, restrictions = "inequalities")
}
\arguments{
\item{x}{a vector with data (for multinomial models) or a vector of counts of successes, or a two-dimensional table
(or matrix) with 2 columns, giving the counts of successes and failures, respectively (for binomial models).}

\item{restrictions}{specifies whether to extract restriction list for \code{equalities} or \code{inequalities}. Default is \code{inequalities}.}
}
\value{
Extracts restriction list and associated hypothesis from an object of class \code{bmult}

Restriction list containing the following elements:
\describe{
\item{\code{$full_model}}{\itemize{
\item  \code{hyp}: character. Vector containing the informed hypothesis as specified by the user
\item  \code{parameters_full}: character. Vector containing the names for each constrained parameter
\item \code{alpha_full}: numeric. Vector containing the concentration parameters
of the Dirichlet distribution (when evaluating ordered multinomial parameters) or alpha parameters of the
beta distribution (when evaluating ordered binomial parameters)
\item  \code{beta_full}: numeric. Vector containing the values of beta parameters of the beta distribution
(when evaluating ordered binomial parameters)
\item  \code{counts_full}: numeric. Vector containing data values (when evaluating multinomial parameters), or number
of successes (when evaluating ordered binomial parameters)
\item  \code{total_full}: numeric. Vector containing the number of observations (when evaluating ordered binomial
parameters, that is, number of successes and failures)
}}
\item{\code{$equality_constraints}}{\itemize{
\item  \code{hyp}: list. Contains all independent equality constrained hypotheses
\item  \code{parameters_equality}: character. Vector containing the names for each equality constrained parameter.
\item \code{equality_hypotheses}: list. Contains the indexes of each equality constrained parameter. Note that these indices are based on the vector of all factor levels
\item  \code{alpha_equalities}: list. Contains the concentration parameters for equality constrained hypotheses (when evaluating multinomial parameters) or alpha parameters of the
beta distribution (when evaluating ordered binomial parameters).
\item  \code{beta_equalities}: list. Contains the values of beta parameters of the beta distribution (when evaluating ordered binomial parameters)
\item  \code{counts_equalities}: list. Contains data values (when evaluating multinomial parameters), or number of successes
(when evaluating ordered binomial parameters) of each equality constrained parameter
\item  \code{total_equalitiesl}: list. Contains the number of observations of each equality constrained parameter (when evaluating ordered binomial
parameters, that is, number of successes and failures)
}}
\item{\code{$inequality_constraints}}{\itemize{
\item  \code{hyp}: list. Contains all independent inequality constrained hypotheses
\item  \code{parameters_inequality}: list. Contains the names for each inequality constrained parameter
\item \code{inequality_hypotheses}: list. Contains the indices of each inequality constrained parameter
\item  \code{alpha_inequalities}: list. Contains for inequality constrained hypotheses the concentration parameters
of the Dirichlet distribution (when evaluating ordered multinomial parameters) or alpha parameters of the beta distribution (when
evaluating ordered binomial parameters).
\item  \code{beta_inequalities}: list. Contains for inequality constrained hypotheses the values of beta parameters of
the beta distribution (when evaluating ordered binomial parameters).
\item  \code{counts_inequalities}: list. Contains for inequality constrained parameter data values (when evaluating
multinomial parameters), or number of successes (when evaluating ordered binomial parameters).
\item  \code{total_inequalities}: list. Contains for each inequality constrained parameter the number of observations
(when evaluating ordered binomial parameters, that is, number of successes and failures).
\item  \code{boundaries}: list that lists for each inequality constrained parameter the index of parameters that
serve as its upper and lower bounds. Note that these indices refer to the collapsed categories (i.e., categories after conditioning
for equality constraints). If a lower or upper bound is missing, for instance because the current parameter is set to be the
smallest or the largest, the bounds take the value \code{int(0)}.
\item  \code{nr_mult_equal}: list. Contains multiplicative elements of collapsed categories
\item  \code{nr_mult_free}: list. Contains multiplicative elements of free parameters
\item  \code{mult_equal}: list. Contains for each lower and upper bound of each inequality constrained parameter
necessary multiplicative elements to recreate the implied order restriction, even for collapsed parameter values. If
there is no upper or lower bound, the multiplicative element will be 0.
\item  \code{nineq_per_hyp}: numeric. Vector containing the total number of inequality constrained parameters
for each independent inequality constrained hypotheses.
\item  \code{direction}: character. Vector containing the direction for each independent inequality constrained
hypothesis. Takes the values \code{smaller} or \code{larger}.
}}
}
}
\description{
Extracts restriction list from an object of class \code{bmult}

Encodes the user specified informed hypothesis. It creates a separate restriction list
for the full model, and all independent equality and inequality constraints. The returned list features relevant
information for the transformation and sampling of the model parameters, such as information about the upper and
lower bound for each parameter, and the indexes of equality constrained and free parameters.
}
\details{
The restriction list can be created for both binomial and multinomial models. If multinomial models are specified,
the arguments \code{b} and \code{n} should be left empty and \code{x} should not be a table or matrix.
}
\examples{
# data
x <- c(3, 4, 10, 11)
n <- c(15, 12, 12, 12)
# priors
a <- c(1, 1, 1, 1)
b <- c(1, 1, 1, 1)
# informed hypothesis
factor_levels <- c('theta1', 'theta2', 'theta3', 'theta4')
Hr            <- c('theta1', '<',  'theta2', '<', 'theta3', '<', 'theta4')

## Multinomial Case
out_mult  <- mult_bf_informed(x=x, Hr=Hr, a=a, factor_levels=factor_levels,
niter=1e3, seed=2020)
generate_restriction_list <- generate_restriction_list(out_mult)
# Restriction list for ordered multinomial
x <- c(1, 4, 1, 10)
a <- c(1, 1, 1, 1)
factor_levels <- c('mult1', 'mult2', 'mult3', 'mult4')
Hr <- c('mult2 > mult1 , mult3 = mult4')
restrictions <- generate_restriction_list(x=x, Hr=Hr, a=a, 
factor_levels=factor_levels)
}
