% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lifestresses.R
\docType{data}
\name{lifestresses}
\alias{lifestresses}
\title{Memory of Life Stresses}
\format{
A \code{data.frame} with 18 rows and 3 variables:
\describe{
\item{month}{The month in which participants reported a stressful life event.}
\item{stress.freq}{The number of participants who reported a life stress in the particular month prior to an interview.}
\item{stress.percentage}{The percentage of participants who reported a life stress in the particular month prior to an interview.}
}
}
\usage{
data(lifestresses)
}
\description{
This data set, "lifestresses", provides the number of reported life stresses (summed across participants)
that occurred in specific months prior to an interview. This data set contains the subset of 147 participants who
reported one negative life event over the time span of 18 months prior to an interview. Description taken from the JASP (2020)
data library.
}
\examples{
data(lifestresses)
# Prior specification 
# We assign a uniform Dirichlet distribution, that is, we set all 
# concentration parameters to 1
a             <- rep(1, 18)
counts        <- lifestresses$stress.freq
factor_levels <- lifestresses$month
# Test the following restricted Hypothesis:
# Hr: month1 > month2 > ... > month18 
Hr            <- paste0(1:18, collapse=">"); Hr
out  <- multBfInformed(factor_levels, Hr, a, counts,  niter=5e4, 
bf_type = 'BFre', seed = 4)
summary(out)
}
\references{
\insertRef{haberman1978}{multibridge}

\insertRef{jasp}{multibridge}

\insertRef{sarafoglou2020evaluatingPreprint}{multibridge}

\insertRef{uhlenhuth1977remembering}{multibridge}
}
\keyword{datasets}
