% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_multBayesInformed.R
\name{multBayesInformed}
\alias{multBayesInformed}
\title{Evaluates Informed Hypotheses on Multinomial Parameters}
\usage{
multBayesInformed(
  factor_levels,
  Hr,
  a,
  counts = NULL,
  niter = 50000,
  bf_type = "LogBFer",
  seed = NULL
)
}
\arguments{
\item{factor_levels}{character. vector with categories}

\item{Hr}{character. vector encoding the user specified order restriction}

\item{a}{numeric. vector with concentration parameters}

\item{counts}{numeric. vector with data}

\item{niter}{numeric. vector with number of samples to be drawn}

\item{bf_type}{'bf_type'. The Bayes factor type can be either 'LogBFer', 'BFer', or 'BFre'. Default is 'LogBFer'}

\item{seed}{set the seed for version control.}
}
\value{
list consisting of the following elements:
(1) BF: Bayes factor for restricted hypothesis compared to the encompassing hypothesis
(2) restrictions: full restriction list
(3) logBFe_equalities: log Bayes factor for equality constrained parameters
(4) logBFe_inequalities: log Bayes factor for inequality constrained parameters
}
\description{
Include a description here.
}
\examples{
# data
counts <- c(3, 4, 10, 11, 7, 30)
# priors
a <- c(1, 1, 1, 1, 1, 1)
# restricted hypothesis
factor_levels <- c('theta1', 'theta2', 'theta3', 'theta4', 'theta5', 'theta6')
Hr            <- c('theta1', '<',  'theta2', '&', 'theta3', '=', 'theta4', ',', 'theta5', '<', 'theta6')
output_total  <- multBayesInformed(factor_levels, Hr, a, counts, niter = 5e3, bf_type = 'LogBFer', seed=2020)
summary(output_total)
}
\references{
Include multinomial paper
}
\seealso{
Other evalInformed: 
\code{\link{binomBayesInformed}()}
}
\concept{evalInformed}
