% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_equalities_mult.R
\name{multBfEquality}
\alias{multBfEquality}
\title{Computes Bayes Factors For Equality Constrained Multinomial Parameters}
\usage{
multBfEquality(x, a, theta = rep(1/length(a), length(a)))
}
\arguments{
\item{x}{numeric. Vector with data}

\item{a}{numeric. Vector with concentration parameters of Dirichlet distribution. Default sets all concentration parameters to 1}

\item{theta}{numeric. Values of interest. Default is 1/K}
}
\value{
list consisting of the following elements:
\describe{
\item{\code{$bf}}{\code{data.frame} containing the Bayes factors \code{LogBFe0}, \code{BFe0}, and \code{BF0e}}
\item{\code{$expected}}{numeric. vector with expected values}
}
}
\description{
Computes Bayes factor for equality constrained multinomial parameters
using the standard Bayesian multinomial test.
Null hypothesis \eqn{H_0} states that category proportions are exactly equal to predefined values.
Alternative hypothesis \eqn{H_e} states that category proportions are free to vary.
}
\examples{
data(lifestresses)
x <- lifestresses$stress.freq
a <- rep(1, nrow(lifestresses))
multBfEquality(x=x, a=a)
}
\seealso{
Other functions to evaluate informed hypotheses: 
\code{\link{binomBfEquality}()},
\code{\link{binomBfInequality}()},
\code{\link{binomBfInformed}()},
\code{\link{multBfInequality}()},
\code{\link{multBfInformed}()}
}
\concept{functions to evaluate informed hypotheses}
