% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_inequalities_mult.R
\name{multBfInequality}
\alias{multBfInequality}
\title{Computes Bayes Factors For Inequality Constrained Multinomial Parameters}
\usage{
multBfInequality(
  samples,
  restrictions,
  alpha = rep(1, ncol(samples)),
  counts = NULL,
  prior = FALSE,
  index = 1,
  maxiter = 1000,
  seed = NULL,
  ...
)
}
\arguments{
\item{samples}{matrix of dimension (nsamples x nparams) with samples from truncated Dirichlet density}

\item{restrictions}{either \code{character vector} containing the user specified order restriction or \code{list} of class \code{bmult_rl} as returned from \code{generateRestrictionList} that encodes
inequality constraints for each independent restriction.}

\item{alpha}{numeric vector with concentration parameters}

\item{counts}{numeric vector with data}

\item{prior}{logical. If TRUE the function will ignore the data and sample from the prior distribution}

\item{index}{index of current restriction. Default is 1.}

\item{maxiter}{maximum number of iterations for the iterative updating scheme. Default is 1,000 to avoid infinite loops.}

\item{seed}{set the seed for version control.}

\item{...}{additional arguments (currently ignored).}
}
\value{
list consisting of the following elements:
(1) eval: list consisting of the following elements
q11: log posterior evaluations for posterior samples.
q12: log proposal evaluations for posterior samples.
q21: log posterior evaluations for samples from proposal.
q22: log proposal evaluations for samples from proposal.
(2) niter: number of iterations of the iterative updating scheme.
(3) logml: estimate of log marginal likelihood.
(4) hyp: character vector that contains the inequality constrained hypothesis
}
\description{
Computes Bayes factor for inequality constrained multinomial parameters using a bridge sampling routine.
Restricted hypothesis Hr states that category proportions follow a particular trend.
Alternative hypothesis He states that category proportions are free to vary.
}
