% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_multBayesInformed.R
\name{multBfInformed}
\alias{multBfInformed}
\title{Evaluates Informed Hypotheses on Multinomial Parameters}
\usage{
multBfInformed(
  x,
  Hr,
  a = rep(1, length(x)),
  factor_levels = NULL,
  cred_level = 0.95,
  niter = 5000,
  bf_type = "LogBFer",
  seed = NULL,
  maxiter = 1000,
  nburnin = niter * 0.05
)
}
\arguments{
\item{x}{numeric. Vector with data}

\item{Hr}{character. Vector encoding the user specified restricted hypothesis. Use either specified factor_levels or indices to refer to parameters}

\item{a}{numeric. Vector with concentration parameters of Dirichlet distribution. Default sets all concentration parameters to 1}

\item{factor_levels}{character. Vector with category names}

\item{cred_level}{numeric. Credible interval for the posterior point estimates. Must be a single number between 0 and 1}

\item{niter}{numeric. Vector with number of samples to be drawn}

\item{bf_type}{character. The Bayes factor type. Ean be either 'LogBFer', 'BFer', or 'BFre'. Default is 'LogBFer'}

\item{seed}{numeric. Sets the seed for version control}

\item{maxiter}{maximum number of iterations for the iterative updating scheme. Default is 1,000 to avoid infinite loops.}

\item{nburnin}{numeric. A single value specifying the number of burn-in samples. Minimum number of burn-in samples is 10. Default is 5\% of the number of
samples. Burn-in samples are removed automatically after the sampling}
}
\value{
List consisting of the following elements
\itemize{
\item \code{BF}: Bayes factor for restricted hypothesis compared to the encompassing hypothesis
\item \code{cred_level}: User specified credible interval
\item \code{restrictions}: full restriction list
\item \code{logBFe_equalities}: log Bayes factor for equality constrained parameters
\item \code{logBFe_inequalities}: log Bayes factor for inequality constrained parameters
}
}
\description{
Evaluates Informed Hypotheses on Multinomial Parameters using bridge sampling.
Restricted hypothesis \eqn{H_r} states that category proportions obey a particular constraint.
Alternative hypothesis \eqn{H_e} states that category proportions are free to vary.
}
\details{
The model assumes that data follow a multinomial distribution and assigns a Dirichlet distribution as prior for the model parameters
(i.e., underlying category proportions). That is:
\deqn{x ~ Multinomial(N, \theta)}
\deqn{\theta ~ Dirichlet(\alpha)}

The following signs can be used to encode restricted hypotheses: \code{"<"} and \code{">"} for inequality constraints, \code{=} for equality constraints,
\code{","} for free parameters, and \code{"&"} for independent hypotheses. The restricted hypothesis can either be a string or a character vector.
For instance, the hypothesis \code{c("theta1 < theta2, theta3")} means
\itemize{
\item \code{theta1} is smaller than both \code{theta2} and \code{theta3}
\item The parameters \code{theta2} and \code{theta3} both have \code{theta1} as lower bound, but are not influenced by each other.
}
The hypothesis \code{c("theta1 < theta2 = theta3 & theta4 > theta5")} means that
\itemize{
\item \code{theta1} is smaller than \code{theta2} and \code{theta3}
\item \code{theta2} and \code{theta3} are assumed to be equal
\item \code{theta4} is larger than \code{theta5}
\item The restrictions on the parameters \code{theta1}, \code{theta2}, and \code{theta3} do
not influence the restrictions on the parameters \code{theta4} and \code{theta5}.
}
}
\examples{
# data
x <- c(3, 4, 10, 11, 7, 30)
# priors
a <- c(1, 1, 1, 1, 1, 1)
# restricted hypothesis
factor_levels <- c('theta1', 'theta2', 'theta3', 'theta4', 'theta5', 'theta6')
Hr            <- c('theta1', '<',  'theta2', '&', 'theta3', '=', 'theta4', ',', 'theta5', '<', 'theta6')
output_total  <- multBfInformed(x, Hr, a, factor_levels, seed=2020)

}
\references{
\insertRef{damien2001sampling}{multibridge}

\insertRef{gronau2017tutorial}{multibridge}

\insertRef{fruhwirth2004estimating}{multibridge}

\insertRef{sarafoglou2020evaluatingPreprint}{multibridge}
}
\seealso{
Other functions to evaluate informed hypotheses: 
\code{\link{binomBfEquality}()},
\code{\link{binomBfInequality}()},
\code{\link{binomBfInformed}()},
\code{\link{multBfEquality}()},
\code{\link{multBfInequality}()}
}
\concept{functions to evaluate informed hypotheses}
