% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_sampling_mult.R
\name{multTruncatedSampling}
\alias{multTruncatedSampling}
\title{Get Samples From Truncated Dirichlet Density}
\usage{
multTruncatedSampling(
  inequalities,
  index = 1,
  niter = 1e+05,
  prior = FALSE,
  nburnin = niter * 0.05,
  seed = NULL
)
}
\arguments{
\item{inequalities}{list that contains inequality constraints for each independent inequality constrained hypotheses.
The list is created in the generateRestrictionList function. In consists of the following elements:}

\item{index}{integer that matches for which inequality constrained hypothesis samples should be drawn. Only relevant
when multiple independent inequality constrained hypotheses are specified. Default is 1.}

\item{niter}{number of samples.}

\item{prior}{logical. If TRUE the function will ignore the data values and sample from the prior distribution instead.}

\item{nburnin}{number of burn-in samples. Minimum number of burn-in samples is 10. Default is 5\% of the number of
samples. Burn-in samples are removed automatically after the sampling.}

\item{seed}{set the seed for version control.}
}
\value{
matrix of dimension niter * nsamples containing prior or posterior samples from truncated Dirichlet distribution.
}
\description{
Get Samples From Truncated Dirichlet Density
}
