% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{print.bmult_bridge}
\alias{print.bmult_bridge}
\title{print method for class \code{bmult_bridge}}
\usage{
\method{print}{bmult_bridge}(x, ...)
}
\arguments{
\item{x}{object of class \code{bmult_bridge} as returned from \code{multBayesBfInequality}}

\item{...}{further arguments, currently ignored}
}
\value{
The print methods print the results and return nothing
}
\description{
print method for class \code{bmult_bridge}
}
