% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{restriction_list.bmult}
\alias{restriction_list.bmult}
\title{Extracts restriction list from an object of class \code{bmult}}
\usage{
\method{restriction_list}{bmult}(x, restrictions = "inequalities")
}
\arguments{
\item{x}{object of class \code{bmult} as returned from \code{multBayesInformed}}
}
\value{
Extracts restriction list and associated hypothesis from an object of class \code{bmult}
}
\description{
Extracts restriction list from an object of class \code{bmult}
}
