% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{restriction_list}
\alias{restriction_list}
\title{S3 method for class 'restriction_list.bmult'}
\usage{
restriction_list(x, restrictions = "inequalities")
}
\description{
S3 method for class 'restriction_list.bmult'
}
