% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{summary.bmult_bridge}
\alias{summary.bmult_bridge}
\title{summary method for class \code{bmult_bridge}}
\usage{
\method{summary}{bmult_bridge}(x)
}
\arguments{
\item{x}{object of class \code{bmult_bridge} as returned from \code{multBayesBfInequality} or \code{binomBayesBfInequality}}
}
\value{
The summary method returns a \code{list} which contains the log marginal likelihood and associated error terms.
}
\description{
summary method for class \code{bmult_bridge}
}
