% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_equalities_binom.R
\name{binomBfEquality}
\alias{binomBfEquality}
\title{Computes Bayes Factors For Equality Constrained Binomial Parameters}
\usage{
binomBfEquality(a, b, counts, total)
}
\arguments{
\item{a}{numeric. vector with concentration parameters}

\item{b}{numeric vector with beta parameters}

\item{counts}{numeric. vector with data}

\item{total}{numeric vector with total number of observations}
}
\value{
list consisting of the following elements:
(1) Bf: dataframe containing the Bayes factor logBFe0, BFe0, BF0e
}
\description{
Computes Bayes factor for equality constrained binomial parameters.
Null hypothesis H0 states that binomial proportions are exactly equal.
Alternative hypothesis He states that binomial proportions are free to vary.
}
