% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{plot.summary.bmult}
\alias{plot.summary.bmult}
\title{Title}
\usage{
\method{plot}{summary.bmult}(x, main = NULL)
}
\arguments{
\item{x}{A \code{bmult}-object returned by \code{summary()}.}

\item{main}{\code{character}. A string used as title. Defaults to the informed
hypothesis and the Bayes factor.}
}
\value{
Invisiblly returns a \code{data.frame} with the plotted estimates.
}
\description{
Title
}
\examples{
# data
x <- c(3, 4, 10, 11, 7, 30)
# priors
a <- c(1, 1, 1, 1, 1, 1)
# restricted hypothesis
factor_levels <- c('theta1', 'theta2', 'theta3', 'theta4', 'theta5', 
                   'theta6')
                   Hr            <- c('theta1', '<',  'theta2', '&', 'theta3', '=', 'theta4', 
                                      ',', 'theta5', '<', 'theta6')
output_total  <- multBfInformed(x, Hr, a, factor_levels, seed=2020, bf_type = "BFer")
plot(summary(output_total))
}
