% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_S3methods.R
\name{summary.bmult}
\alias{summary.bmult}
\title{summary method for class \code{bmult}}
\usage{
\method{summary}{bmult}(x)
}
\arguments{
\item{x}{object of class \code{bmult} as returned from \code{multBayesInformed}}
}
\value{
The summary method returns a \code{list} which contains the Bayes factor and associated hypotheses for the full
model, but also the separate for the independent equality and inequality constraints.
}
\description{
summary method for class \code{bmult}
}
